+++ src/Core/AccountUnfreezer.php
@@ -3,7 +3,10 @@
 namespace Nadybot\Core;
 
 use function Amp\call;
+use function Safe\json_decode;
+
 use Amp\Http\Client\Connection\{DefaultConnectionFactory, UnlimitedConnectionPool};
+use Amp\Http\Client\Interceptor\SetRequestHeader;
 use Amp\Http\Client\{HttpClient, HttpClientBuilder, Request, Response, SocketException, TimeoutException};
 use Amp\Http\Tunnel\Http1TunnelConnector;
 use Amp\Socket\SocketAddress;
@@ -11,12 +14,18 @@ use Amp\{CancelledException, Loop, Promise, TimeoutCancellationToken};
 use Generator;
 
 use Nadybot\Core\Attributes as NCA;
+use Safe\Exceptions\JsonException;
 use Throwable;
 
 #[NCA\Instance]
 class AccountUnfreezer {
-	public const LOGIN_URL = "https://register.funcom.com/account";
-	public const UNFREEZE_URL = "https://register.funcom.com/account/subscription/ctrl/anarchy/%s/reactivate";
+	public const LOGIN_URL = "https://account.anarchy-online.com/";
+	public const ACCOUNT_URL = "https://account.anarchy-online.com/account/";
+	public const SUBSCRIPTION_URL = "https://account.anarchy-online.com/subscription/%s";
+	public const UNFREEZE_URL = "https://account.anarchy-online.com/uncancel_sub";
+	public const LOGOUT_URL = "https://account.anarchy-online.com/log_out";
+
+	public const DEFAULT_UA = "Mozilla/5.0 (X11; Linux x86_64; rv:109.0) Gecko/20100101 Firefox/110.0";
 
 	private const UNFREEZE_FAILURE = 0;
 	private const UNFREEZE_SUCCESS = 1;
@@ -24,6 +33,8 @@ class AccountUnfreezer {
 
 	private const PROXY_HOST = 'proxy.nadybot.org';
 	private const PROXY_PORT = 22222;
+
+	protected ?string $userAgent = null;
 
 	#[NCA\Logger]
 	private LoggerWrapper $logger;
@@ -34,17 +45,19 @@ class AccountUnfreezer {
 	#[NCA\Inject]
 	private HttpClientBuilder $http;
 
-	public function unfreeze(): bool {
+	public function unfreeze(?int $subscriptionId): bool {
 		$result = false;
-		Loop::run(function () use (&$result): Generator {
+		Loop::run(function () use (&$result, $subscriptionId): Generator {
 			$this->logger->warning('Account frozen, trying to unfreeze');
 
+			/** @var HttpClient */
+			$client = yield $this->getUnfreezeClient();
+
 			do {
-				$client = $this->getUnfreezeClient();
 				$lastResult = self::UNFREEZE_TEMP_ERROR;
 				$proxyText = $this->config->autoUnfreezeUseNadyproxy ? "Proxy" : "Unfreezing";
 				try {
-					$lastResult = yield $this->unfreezeWithClient($client);
+					$lastResult = yield $this->unfreezeWithClient($client, $subscriptionId);
 				} catch (CancelledException) {
 					$this->logger->notice("{$proxyText} not working or too slow. Retrying.");
 				} catch (SocketException) {
@@ -67,36 +80,22 @@ class AccountUnfreezer {
 		return $result;
 	}
 
-	/** @return Promise<int> */
-	protected function unfreezeWithClient(
-		HttpClient $client,
-	): Promise {
+	/** @return Promise <string> */
+	protected function getSessionCookie(HttpClient $client): Promise {
 		return call(function () use ($client): Generator {
-			$login = strtolower($this->config->login);
-			$user = strtolower($this->config->autoUnfreezeLogin ?? $login);
-			$password = $this->config->password;
-			$request = new Request(self::LOGIN_URL, "POST");
-			$request->setBody(http_build_query([
-				"__ac_name" => $login,
-				"__ac_password" => $password,
-			]));
-			$request->addHeader("Content-Type", "application/x-www-form-urlencoded");
-			$request->addHeader("Referer", self::LOGIN_URL);
+			$request = new Request(self::LOGIN_URL, "GET");
 			$request->setTcpConnectTimeout(5000);
 			$request->setTlsHandshakeTimeout(5000);
 			$request->setTransferTimeout(5000);
 
 			/** @var Response */
-			$response = yield $client->request(
-				$request,
-				new TimeoutCancellationToken(10000)
-			);
+			$response = yield $client->request($request, new TimeoutCancellationToken(10000));
 
-			if ($response->getStatus() !== 302) {
-				$this->logger->error('Unable to login to the account management website: {code}', [
+			if ($response->getStatus() !== 200) {
+				$this->logger->error('Unable to login to get session cookie: {code}', [
 					"code" => $response->getStatus(),
 				]);
-				return self::UNFREEZE_TEMP_ERROR;
+				throw new UnfreezeTmpException();
 			}
 			$cookies = $response->getHeaderArray('Set-Cookie');
 			$cookieValues = [];
@@ -106,52 +105,236 @@ class AccountUnfreezer {
 					$cookieValues []= $cookieParts[0];
 				}
 			}
-
-			$request = new Request(sprintf(self::UNFREEZE_URL, $user), "POST");
-			$request->setBody(http_build_query(["process" => 'submit']));
-			$request->addHeader("Content-Type", "application/x-www-form-urlencoded");
-			$request->addHeader("Referer", sprintf(self::UNFREEZE_URL, $user));
-			$request->addHeader("Cookie", implode("; ", $cookieValues));
-
-			/** @var Response */
-			$response = yield $client->request($request);
-			if ($response->getStatus() === 500) {
-				$this->logger->warning("There was an error unfreezing the account");
-				return self::UNFREEZE_TEMP_ERROR;
-			}
-			if ($response->getStatus() !== 200) {
-				$this->logger->warning("There was an error unfreezing the account");
-				return self::UNFREEZE_FAILURE;
-			}
-			$body = yield $response->getBody()->buffer();
-			if (strpos($body, "<div>Subscription Reactivated</div>") !== false) {
-				return self::UNFREEZE_SUCCESS;
-			}
-			if (strpos($body, "<div>This account is not cancelled or frozen</div>") !== false) {
-				$this->logger->notice("According to Funcom, the account isn't frozen.");
-				return self::UNFREEZE_SUCCESS;
-			}
-			$this->logger->warning("There was an error unfreezing the account");
-			return self::UNFREEZE_FAILURE;
+			return join("; ", $cookieValues);
 		});
 	}
 
-	/** Get a HttpClient that uses the Nadybot proxy to unfreeze an account */
-	private function getUnfreezeClient(): HttpClient {
-		if ($this->config->autoUnfreezeUseNadyproxy === false) {
-			return $this->http->followRedirects(0)->build();
-		}
-		return $this->http
-			->followRedirects(0)
-			->usingPool(
-				new UnlimitedConnectionPool(
-					new DefaultConnectionFactory(
-						new Http1TunnelConnector(
-							new SocketAddress(self::PROXY_HOST, self::PROXY_PORT)
+	/** @return Promise<void> */
+	protected function loginToAccount(HttpClient $client, string $cookie): Promise {
+		return call(function () use ($client, $cookie): Generator {
+			$login = strtolower($this->config->login);
+			$user = strtolower($this->config->autoUnfreezeLogin ?? $login);
+			$password = $this->config->autoUnfreezePassword ?? $this->config->password;
+			$request = new Request(self::LOGIN_URL, "POST");
+			$request->setBody(http_build_query([
+				"nickname" => $user,
+				"password" => $password,
+			]));
+			$request->addHeader("Cookie", $cookie);
+			$request->addHeader("Content-Type", "application/x-www-form-urlencoded");
+			$request->addHeader("Referer", self::LOGIN_URL);
+			$request->setTcpConnectTimeout(5000);
+			$request->setTlsHandshakeTimeout(5000);
+			$request->setTransferTimeout(5000);
+
+			/** @var Response */
+			$response = yield $client->request($request, new TimeoutCancellationToken(10000));
+
+			if ($response->getStatus() !== 302) {
+				$errorMsg = "HTTP-Code " . $response->getStatus();
+				if ($response->getStatus() === 200) {
+					$body = yield $response->getBody()->buffer();
+					if (preg_match('/<div class="alert alert-danger">(.+?)<\/div>/s', $body, $matches)) {
+						$errorMsg = trim(strip_tags($matches[1]));
+					}
+				}
+				$this->logger->error('Unable to login to the account management website: {error}', [
+					"error" => $errorMsg,
+				]);
+				throw new UnfreezeTmpException();
+			}
+		});
+	}
+
+	/** @return Promise<void> */
+	protected function switchToAccount(HttpClient $client, string $cookie, int $accountId): Promise {
+		return call(function () use ($client, $cookie, $accountId): Generator {
+			$request = new Request(sprintf(self::SUBSCRIPTION_URL, $accountId), "GET");
+			$request->addHeader("Cookie", $cookie);
+			$request->addHeader("Referer", self::LOGIN_URL);
+			$request->setTcpConnectTimeout(5000);
+			$request->setTlsHandshakeTimeout(5000);
+			$request->setTransferTimeout(5000);
+
+			/** @var Response */
+			$response = yield $client->request($request, new TimeoutCancellationToken(10000));
+
+			if ($response->getStatus() !== 302) {
+				$this->logger->error('Unable to switch to the correct account: {code}', [
+					"code" => $response->getStatus(),
+				]);
+				throw new UnfreezeTmpException();
+			}
+		});
+	}
+
+	/** @return Promise<string> */
+	protected function loadAccountPage(HttpClient $client, string $cookie): Promise {
+		return call(function () use ($client, $cookie): Generator {
+			$request = new Request(self::ACCOUNT_URL, "GET");
+			$request->addHeader("Cookie", $cookie);
+			$request->addHeader("Referer", self::LOGIN_URL);
+			$request->setTcpConnectTimeout(5000);
+			$request->setTlsHandshakeTimeout(5000);
+			$request->setTransferTimeout(5000);
+
+			/** @var Response */
+			$response = yield $client->request($request, new TimeoutCancellationToken(10000));
+
+			if ($response->getStatus() !== 200) {
+				$this->logger->error('Unable to read account page: {code}', [
+					"code" => $response->getStatus(),
+				]);
+				throw new UnfreezeTmpException();
+			}
+			return $response->getBody()->buffer();
+		});
+	}
+
+	/** @return Promise<void> */
+	protected function uncancelSub(HttpClient $client, string $cookie): Promise {
+		return call(function () use ($client, $cookie): Generator {
+			$request = new Request(self::UNFREEZE_URL, "GET");
+			$request->addHeader("Cookie", $cookie);
+			$request->addHeader("Referer", self::ACCOUNT_URL);
+			$request->setTcpConnectTimeout(5000);
+			$request->setTlsHandshakeTimeout(5000);
+			$request->setTransferTimeout(5000);
+
+			/** @var Response */
+			$response = yield $client->request($request, new TimeoutCancellationToken(10000));
+
+			if ($response->getStatus() !== 302) {
+				$this->logger->error('Unable to unfreeze account: {code}', [
+					"code" => $response->getStatus(),
+				]);
+				throw new UnfreezeTmpException();
+			}
+		});
+	}
+
+	/** @return Promise<int> */
+	protected function getSubscriptionId(HttpClient $client, string $cookie): Promise {
+		return call(function () use ($client, $cookie): Generator {
+			/** @var string */
+			$body = yield $this->loadAccountPage($client, $cookie);
+			$login = strtolower($this->config->login);
+			if (!preg_match(
+				'/<li><a href="\/subscription\/(\d+)">' . preg_quote($login, '/') . '<\/a><\/li>/s',
+				$body,
+				$matches
+			)) {
+				throw new UnfreezeFatalException("Account {$login} not on this login.");
+			}
+			return (int)$matches[1];
+		});
+	}
+
+	/** @return Promise <string> */
+	protected function logout(HttpClient $client, string $cookie): Promise {
+		return call(function () use ($client, $cookie): Generator {
+			$request = new Request(self::LOGOUT_URL, "GET");
+			$request->addHeader("Cookie", $cookie);
+			$request->addHeader("Referer", self::ACCOUNT_URL);
+			$request->setTcpConnectTimeout(5000);
+			$request->setTlsHandshakeTimeout(5000);
+			$request->setTransferTimeout(5000);
+
+			/** @var Response */
+			$response = yield $client->request($request, new TimeoutCancellationToken(10000));
+
+			if ($response->getStatus() !== 302) {
+				$this->logger->error('Error logging out: {code}', [
+					"code" => $response->getStatus(),
+				]);
+			}
+		});
+	}
+
+	/** @return Promise<int> */
+	protected function unfreezeWithClient(
+		HttpClient $client,
+		?int $subscriptionId
+	): Promise {
+		return call(function () use ($client, $subscriptionId): Generator {
+			try {
+				$sessionCookie = yield $this->getSessionCookie($client);
+				yield $this->loginToAccount($client, $sessionCookie);
+
+				/** @var int */
+				$accountId = yield $this->getSubscriptionId($client, $sessionCookie);
+				if (isset($subscriptionId) && $accountId !== $subscriptionId) {
+					$this->logger->error("Subscription {subscription} is not managed via given login.", [
+						"subscription" => $subscriptionId,
+					]);
+					return self::UNFREEZE_FAILURE;
+				}
+				yield $this->switchToAccount($client, $sessionCookie, $accountId);
+				$mainBody = yield $this->loadAccountPage($client, $sessionCookie);
+				if (!str_contains($mainBody, "Free Account")) {
+					$this->logger->error("Refusing to unfreeze a paid account");
+					return self::UNFREEZE_FAILURE;
+				}
+				yield $this->uncancelSub($client, $sessionCookie);
+				yield $this->logout($client, $sessionCookie);
+				return self::UNFREEZE_SUCCESS;
+			} catch (UnfreezeTmpException) {
+				return self::UNFREEZE_TEMP_ERROR;
+			} catch (UnfreezeFatalException) {
+				return self::UNFREEZE_FAILURE;
+			}
+		});
+	}
+
+	/** @return Promise<?string> */
+	protected function getUserAgent(): Promise {
+		return call(function (): Generator {
+			$this->logger->info("Getting most popular user agent");
+			$client = $this->http->build();
+			$request = new Request("https://raw.githubusercontent.com/Kikobeats/top-user-agents/master/index.json");
+
+			/** @var Response */
+			$response = yield $client->request($request);
+			if ($response->getStatus() !== 200) {
+				return null;
+			}
+			$body = yield $response->getBody()->buffer();
+			try {
+				$json = json_decode($body, false);
+				if (!is_array($json) || !isset($json[0]) || !is_string($json[0])) {
+					return null;
+				}
+			} catch (JsonException) {
+				return null;
+			}
+			return $json[0];
+		});
+	}
+
+	/**
+	 * Get a HttpClient that uses the Nadybot proxy to unfreeze an account
+	 *
+	 * @return Promise<HttpClient>
+	 */
+	private function getUnfreezeClient(): Promise {
+		return call(function (): Generator {
+			$this->userAgent ??= yield $this->getUserAgent();
+			$this->userAgent ??= self::DEFAULT_UA;
+			$this->logger->info("Using user agent {agent}", ["agent" => $this->userAgent]);
+			$builder = $this->http->followRedirects(0)
+					->intercept(new SetRequestHeader("User-Agent", $this->userAgent));
+			if ($this->config->autoUnfreezeUseNadyproxy !== false) {
+				$builder = $builder->usingPool(
+					new UnlimitedConnectionPool(
+						new DefaultConnectionFactory(
+							new Http1TunnelConnector(
+								new SocketAddress(self::PROXY_HOST, self::PROXY_PORT)
+							)
 						)
 					)
-				)
-			)
-			->build();
+				);
+			}
+			return $builder->build();
+		});
 	}
 }